fix models

+++ backend/models/Ingrediente.js
@@ -2,7 +2,7 @@ const Ingrediente = (sequelize,DataTypes) => {
     return sequelize.define('Ingrediente', {
         nome: DataTypes.STRING,
         unidade_medida: DataTypes.STRING,
-        preco: DataTypes.DOUBLE
+        preco: DataTypes.FLOAT
     },{
         tableName: 'Ingrediente'
     });
